refactor: 서비스 간 의존하는 구조 수정 (#37)
